fix cache/memory node pruning
